refactor & feat: merge interface, add more validator
- merge interface to the model file to avoid more imports
- add password validator
- add role validator
